WORK IN PROGRESS: A raw attempt at a columnar segment builder.
Need to do some performance measures and see if this even helps.
Some parts can be used towards unified segment creation.

The idea is that for realtime segment builds, we can:

   get sorted dictionary ids (essentially new to old dictinary id mapping)
   get/put the values to create new dictionary.
   use only dictionary IDs to create fwd index and inverted index.
   Use the raw value readers we already have to create raw index.

We need to check if this is a significant performance gain before deciding
how to set up the classes correctly so that we don't have random dependencies
everywhere

Things that don't work the same as before:
  Min/Max value of time column not set in metadata.properties
  DefaultNull for boolean columns is set to false instead of null.
Things that probably won't work in this impl:
  Offline segment creation
  starttree segment creation.