Adding a timer calls for serialization of results
We need to measure this, but it is a chicken and egg problem.
We need to somehow time the serialization and then go back and update the
time set in the serialized datatable.
Alternative is to keep some serialization measurements, and add them
in from previous queries.
For now, just adding a timer so that we compute the overhead correctly.